// lib/api/generated/model/requirementOpenIssueRead.ts
/**
 * Generated by orval v8.10.0 🍺
 * Do not edit manually.
 * Syncnesto API
 * OpenAPI spec version: 0.1.0
 */

/**
 * 未決事項読み取り時に返すschema。
 */
export interface RequirementOpenIssueRead {
  related_requirement_id?: number | null;
  issue_code: string;
  title: string;
  description?: string | null;
  impact_scope?: string | null;
  assignee_id?: number | null;
  due_date?: string | null;
  status?: string;
  resolution?: string | null;
  id: number;
  document_id: number;
  version: number;
  created_by?: number | null;
  updated_by?: number | null;
  created_at: string;
  updated_at: string;
}